[Mosaic] Add a new kernel-wide compiler flag `strict_numerics` to pallas/mosaic. Default is true, if set to be false, some optimizations will be applied, but no longer provide guarantees on strict numerics.
Also add the first optimization controlled by this flag, which is to increase the number of partial accumulators used in multi_reduction to expose more ILP. The cost is, the reduction order might be changed, leading to potential small numerical changes.

In future, more optimizations could be added under this flag.

PiperOrigin-RevId: 808807691