[Mosaic] Use small second minor tiling for packed type reshape of contiguous vreg slice.
This can be done by mixed unpacking/packing to halve/double sublane tiling. Alternatively, we can unpack to 32 bit and use sublane shuffle, but we lack of support for general tpu::SublaneShuffleOp. And it's not going to be faster IMO as arbitrary shuffle pattern needs to be emulated by store + load.

PiperOrigin-RevId: 809273967